% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambhtmx.R
\name{ambhtmx_app}
\alias{ambhtmx_app}
\title{Creating an ambhtmx app}
\usage{
ambhtmx_app(
  dbname = NULL,
  value = tibble(),
  host = "127.0.0.1",
  port = "8000",
  live = FALSE,
  renderer = NULL
)
}
\description{
Creating an ambhtmx app
}
