% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-amb.R
\name{amb_card}
\alias{amb_card}
\alias{amb_input_text}
\alias{amb_button}
\title{Generate component}
\usage{
amb_card(
  ...,
  class = NULL,
  title = NULL,
  title_icon = NULL,
  title_class = NULL
)

amb_input_text(
  ...,
  id,
  label = NULL,
  value = "",
  input_class = NULL,
  hx_post = NULL
)

amb_button(..., class = "rounded-1", type = "button")
}
\arguments{
\item{...}{htmlobjects to add}

\item{class}{customize}

\item{title}{to customeize the title text}

\item{title_icon}{to customize the title icon}

\item{title_class}{to add more classes to the title}

\item{id}{for the label and the input}

\item{label}{customize}

\item{value}{customize}

\item{input_class}{customize}

\item{hx_post}{customize}

\item{type}{customize}
}
\description{
Generate component
}
\keyword{components}
