% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambhtmx.R
\name{process_loggedin_middleware}
\alias{process_loggedin_middleware}
\title{Process loggedin middleware}
\usage{
process_loggedin_middleware(
  req,
  res,
  user = Sys.getenv("AMBHTMX_USER"),
  cookie_loggedin = "loggedin"
)
}
\description{
Process loggedin middleware
}
