% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambhtmx.R
\name{process_loggedin_redirect}
\alias{process_loggedin_redirect}
\title{Process loggedin redirect}
\usage{
process_loggedin_redirect(
  req,
  res,
  user = Sys.getenv("AMBHTMX_USER"),
  login_url = "/login"
)
}
\description{
Process loggedin redirect
}
