% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambhtmx.R
\name{process_login_get}
\alias{process_login_get}
\title{Render and send login page response}
\usage{
process_login_get(
  req,
  res,
  page_title = "Login",
  main = NULL,
  id = "login_form",
  login_url = "/login",
  style = "margin: 20px",
  cookie_errors = "errors"
)
}
\arguments{
\item{req}{request object}

\item{res}{response object}

\item{page_title}{if you need to customize the title of the page}

\item{main}{if you need to customize the body of the login page}

\item{id}{if you need to customize the id of the login form}

\item{login_url}{if you need to customize the url of the login form}

\item{style}{if you need to customize the styles of the login form}

\item{cookie_errors}{if you need to customize the name of the errors cookie}
}
\value{
the login page response
}
\description{
Render and send login page response
}
