% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambhtmx.R
\name{process_login_post}
\alias{process_login_post}
\title{Process login requests}
\usage{
process_login_post(
  req,
  res,
  user_param = "user",
  password_param = "password",
  user = Sys.getenv("AMBHTMX_USER"),
  password = Sys.getenv("AMBHTMX_PASSWORD"),
  user_error = "Invalid user",
  password_error = "Invalid password",
  cookie_loggedin = "loggedin",
  cookie_errrors = "errors",
  login_url = "/login",
  success_url = "/"
)
}
\description{
Process login requests
}
