% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambhtmx.R
\name{process_logout_get}
\alias{process_logout_get}
\title{Process logout requests}
\usage{
process_logout_get(req, res, cookie_loggedin = "loggedin", success_url = "/")
}
\description{
Process logout requests
}
