% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambhtmx.R
\name{render_page}
\alias{render_page}
\title{Render a custom page with a custom title and main content}
\usage{
render_page(page_title = NULL, main = NULL)
}
\description{
Render a custom page with a custom title and main content
}
