% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambhtmx.R
\name{render_plot}
\alias{render_plot}
\title{Render imatge or ggplot to image tag}
\usage{
render_plot(p)
}
\arguments{
\item{p}{a ggplot or another object that can be printed and captured as a png image}
}
\value{
img htmltools tag with a data encoded src attribute
}
\description{
Render imatge or ggplot to image tag
}
