% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rendering.R
\name{render_png}
\alias{render_png}
\title{Render a png image to a img tag}
\usage{
render_png(p)
}
\arguments{
\item{p}{a ggplot or another object that can be printed and captured as a png image}
}
\value{
img htmltools tag with a data encoded src attribute
}
\description{
Render a png image to a img tag
}
\keyword{rendering}
