% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambhtmx.R
\name{render_tags}
\alias{render_tags}
\title{Render tags to character vector}
\usage{
render_tags(...)
}
\arguments{
\item{...}{one or more htmltools objects.}
}
\value{
a character representation of input
}
\description{
Render tags to character vector
}
