% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-tags.R
\name{script_from_js_tpl}
\alias{script_from_js_tpl}
\title{Generate script from js template}
\usage{
script_from_js_tpl(file, ...)
}
\arguments{
\item{file}{path to a js file}

\item{...}{mutiple named arguments with the value to replace}
}
\description{
Generate script from js template
}
\examples{
if (FALSE){
  # replaces {init} to 0
  script_from_js_tpl("script.js", init = "init")
}
}
\keyword{components}
