% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-tags.R
\name{script_tpl}
\alias{script_tpl}
\title{Generate script from raw js template}
\usage{
script_tpl(raw_content, ...)
}
\arguments{
\item{raw_content}{path to a js file}

\item{...}{mutiple named arguments with the value to replace}
}
\description{
Generate script from raw js template
}
\keyword{components}
