% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambhtmx.R
\name{send_page}
\alias{send_page}
\title{Render a custom page with a custom title and main content}
\usage{
send_page(main, res, ...)
}
\arguments{
\item{main}{htmltools object of the body of the html page}

\item{res}{response object}

\item{...}{other paramters to the render page function}
}
\description{
Render a custom page with a custom title and main content
}
