% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{send_tags}
\alias{send_tags}
\title{Render a custom page with a custom title and main content}
\usage{
send_tags(main = NULL, res, ...)
}
\arguments{
\item{main}{htmltools object to render}

\item{res}{response object}

\item{...}{htmltools object to render}
}
\description{
Render a custom page with a custom title and main content
}
\keyword{render}
