% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-tags.R
\name{style_tpl}
\alias{style_tpl}
\title{Generate style from raw css}
\usage{
style_tpl(raw_content, ...)
}
\arguments{
\item{raw_content}{contents with tpl in css}

\item{...}{mutiple named arguments with the value to replaces}
}
\description{
Generate style from raw css
}
\keyword{components}
