% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-tags.R
\name{tags}
\alias{tags}
\alias{button}
\alias{textarea}
\alias{input}
\alias{label}
\alias{nav}
\alias{li}
\alias{ul}
\alias{ol}
\alias{form}
\alias{style}
\alias{script}
\title{Create HTML tags without requiring htmltools::tags$}
\usage{
button(..., .noWS = NULL, .renderHook = NULL)

textarea(..., .noWS = NULL, .renderHook = NULL)

input(..., .noWS = NULL, .renderHook = NULL)

label(..., .noWS = NULL, .renderHook = NULL)

nav(..., .noWS = NULL, .renderHook = NULL)

li(..., .noWS = NULL, .renderHook = NULL)

ul(..., .noWS = NULL, .renderHook = NULL)

ol(..., .noWS = NULL, .renderHook = NULL)

form(..., .noWS = NULL, .renderHook = NULL)

style(..., .noWS = NULL, .renderHook = NULL)

script(..., .noWS = NULL, .renderHook = NULL)
}
\arguments{
\item{...}{Tag attributes (named arguments) and children (unnamed arguments).
A named argument with an \code{NA} value is rendered as a boolean attributes
(see example). Children may include any combination of:
\itemize{
\item Other tags objects
\item \code{\link[=HTML]{HTML()}} strings
\item \code{\link[=htmlDependency]{htmlDependency()}}s
\item Single-element atomic vectors
\item \code{list()}s containing any combination of the above
}}

\item{.noWS}{Character vector used to omit some of the whitespace that would
normally be written around this tag. Valid options include \code{before},
\code{after}, \code{outside}, \code{after-begin}, and \code{before-end}.
Any number of these options can be specified.}

\item{.renderHook}{A function (or list of functions) to call when the \code{tag} is rendered. This
function should have at least one argument (the \code{tag}) and return anything
that can be converted into tags via \code{\link[=as.tags]{as.tags()}}. Additional hooks may also be
added to a particular \code{tag} via \code{\link[=tagAddRenderHook]{tagAddRenderHook()}}.}
}
\value{
A \code{list()} with a \code{shiny.tag} class that can be converted into an
HTML string via \code{as.character()} and saved to a file with \code{save_html()}.
}
\description{
Create an R object that represents an HTML tag. For convenience, common HTML
tags (e.g., \verb{<div>}) can be created by calling for their tag name directly
(e.g., \code{div()}). To create less common HTML5 (or SVG) tags (e.g.,
\verb{<article>}), use the \code{tags} list collection (e.g., \code{tags$article()}). To
create other non HTML/SVG tags, use the lower-level \code{tag()} constructor.
}
\keyword{components}
