% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambhtmx.R
\name{ambhtmx_app}
\alias{ambhtmx_app}
\title{Creating an ambiorix + htmx app}
\usage{
ambhtmx_app(
  dbname = NULL,
  value = tibble::tibble(),
  host = "127.0.0.1",
  port = "8000",
  live = "",
  render_index = NULL,
  render_row = NULL,
  favicon = NULL
)
}
\description{
Creating an ambiorix + htmx app
}
