% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambhtmx.R
\name{process_error_post}
\alias{process_error_post}
\title{Process error post requests}
\usage{
process_error_post(
  req,
  res,
  errors = NULL,
  cookie_errrors = "errors",
  error_url = NULL
)
}
\description{
Process error post requests
}
