% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{process_loggedin_middleware}
\alias{process_loggedin_middleware}
\title{Process loggedin middleware}
\usage{
process_loggedin_middleware(
  req,
  res,
  user = Sys.getenv("AMBHTMX_USER"),
  cookie_loggedin = "loggedin"
)
}
\arguments{
\item{req}{request object}

\item{res}{response object}

\item{user}{if you want to customize the required user or it uses AMBHTMX_USER}

\item{cookie_loggedin}{if you need to customize the name of the loggedin cookie}
}
\value{
the updated request with the req$loggedin status
}
\description{
Process loggedin middleware
}
\keyword{auth}
