% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambhtmx.R
\name{process_login_get}
\alias{process_login_get}
\title{Render login forms}
\usage{
process_login_get(
  req,
  res,
  ...,
  page_title = "Login",
  main = NULL,
  id = "login_form",
  login_url = "/login",
  style = "margin: 20px",
  cookie_errors = "errors"
)
}
\description{
Render login forms
}
