% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambhtmx.R
\name{process_logout_get}
\alias{process_logout_get}
\title{Process logout requests}
\usage{
process_logout_get(req, res, cookie_loggedin = "loggedin", success_url = "/")
}
\arguments{
\item{req}{request object}

\item{res}{response object}

\item{cookie_loggedin}{if you need to customize the name of the loggedin cookie}

\item{success_url}{if you need to customize the url of the success loggedin process}
}
\value{
the logout process response
}
\description{
Process logout requests
}
