% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rendering.R
\name{send_page}
\alias{send_page}
\title{Render a page and send the respose}
\usage{
send_page(main = NULL, res, ...)
}
\arguments{
\item{main}{htmltools object of the body of the html page}

\item{res}{response object}

\item{...}{other paramters to the render page function}
}
\description{
Render a page and send the respose
}
\keyword{rendering}
