% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-tags.R
\name{style_from_css_tpl}
\alias{style_from_css_tpl}
\title{Generate style from css template}
\usage{
style_from_css_tpl(file, ...)
}
\arguments{
\item{file}{path to a js file}

\item{...}{mutiple named arguments with the value to replaces}
}
\description{
Generate style from css template
}
\examples{
if (FALSE){
  # replaces "var(--tpl-background)" to "red"
  style_from_css_tpl("styles.css", background = "red")
}
}
